Updated to handle a longer int intake.
Changed function to take in a "const long long unsigned int", instead of a "const long unsigned int"

+++ ft_put_uint.c
@@ -12,7 +12,7 @@
 
 #include "libft/libft.h"
 
-size_t	ft_put_uint(const long unsigned int num)
+size_t	ft_put_uint(const long long unsigned int num)
 {
 	size_t	len;
